backend: authenticated routes

+++ backend/src/routes/memories.ts
@@ -4,8 +4,17 @@ import { z } from 'zod'
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    // verify jwt for all routes
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+
+    // request to list memories by userId to logged user
+    app.get('/memories', async (request) => {
         const memories  = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             },
@@ -20,7 +29,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    // request to list a memorie by memoryId to logged user.
+    app.get('/memories/:id', async (request, reply) => {
         // validation with zod
         const paramsSchema = z.object({
             id: z.string().uuid(),
@@ -36,10 +46,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
             },
         })
 
+        // check if the memory belongs to the logged in user and the memory is not public
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
+
         //  return memory
         return memory
     })
 
+    // request to create a memory by userId to logged user
     app.post('/memories', async (request) => {
         // validate body request with zod
         const bodySchema = z.object({
@@ -57,7 +73,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: 'c227cac7-f972-4ebc-b30a-61b1695057ef'
+                userId: request.user.sub
             },
         })
 
@@ -65,7 +81,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.put('/memories/:id', async (request) => {
+    // request to update a memory by memoryId to logged user
+    app.put('/memories/:id', async (request, reply) => {
         // validate request params with zod
         const paramsSchema = z.object({
             id: z.string().uuid(),
@@ -84,6 +101,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
         // payload with zod validation
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
+        // find the memory
+        const memory =  await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        // check if the memory belongs to the logged in user
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
+
         // update data on database
         const memoryUpdated = await prisma.memory.update({
             where: {
@@ -100,7 +129,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memoryUpdated
     })
 
-    app.delete('/memories/:id', async (request) => {
+    // request to delete a memory by memoryId to logged user
+    app.delete('/memories/:id', async (request, reply) => {
         // validate request params with zod
         const paramsSchema = z.object({
             id: z.string().uuid(),
@@ -108,6 +138,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         // request param with the zod validation
         const { id } = paramsSchema.parse(request.params)
+
+        // find the memory
+        const memory =  await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        // check if the memory belongs to the logged in user
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
 
         // delete specific memory via id
         await prisma.memory.delete({
